requirements.txt
alembic==1.18.4
blinker==1.9.0
certifi==2026.2.25
chardet==7.0.1
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
Flask==3.1.3
flask-cors==6.0.2
gunicorn==25.1.0
idna==3.11
ijson==3.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
packaging==26.0
psycopg2-binary==2.9.11
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.32.5
SQLAlchemy==1.4.24
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
